+++ torchtune/utils/__init__.py
@@ -23,7 +23,7 @@ from ._distributed import (  # noqa
     wrap_fsdp,
 )
 from ._generation import generate  # noqa
-from ._perf_profiler import perf_profiler
+from ._profiler import profiler
 from .argparse import TuneRecipeArgumentParser
 from .checkpointable_dataloader import CheckpointableDataLoader
 from .collate import padded_collate
@@ -81,5 +81,5 @@ __all__ = [
     "OptimizerInBackwardWrapper",
     "create_optim_in_bwd_wrapper",
     "register_optim_in_bwd_hooks",
-    "perf_profiler",
+    "profiler",
 ]
